+++ kaithem/src/sound.py
@@ -8,6 +8,19 @@ from icemedia.sound_player import *  # noqa
 from . import config, directories, messagebus, util
 
 
+def test(output=None):
+    t = "test_" + str(time.time())
+    play_sound("alert.ogg", output=output, handle=t)
+    for i in range(100):
+        if is_playing(t, refresh=True):
+            return
+        time.sleep(0.01)
+    raise RuntimeError("Sound did not report as playing within 1000ms")
+
+
+sound.test = test
+
+
 def init():
     # Todo move sound init to it's own thing
     if "dummy" in sound.backend.backendname.lower():
@@ -15,6 +28,12 @@ def init():
             "/system/notifications/errors",
             "Using a dummy sound backend. Suggest installing MPV if you want sound",
         )
+
+    def special_resolver(fn):
+        if fn == "alert.ogg":
+            return os.path.join(directories.datadir, "static/sounds/72127__kizilsungur__sweetalertsound3.opus")
+        if fn == "error.ogg":
+            return os.path.join(directories.datadir, "static/sounds/423166__plasterbrain__minimalist-sci-fi-ui-error.opus")
 
     def resolver(fn):
         for i in os.listdir(os.path.join(directories.vardir, "modules", "data")):
@@ -24,6 +43,7 @@ def init():
                 break
 
     sound.media_resolvers["kaithem_module"] = resolver
+    sound.media_resolvers["kaithem_special"] = special_resolver
 
     sound.media_paths.append(os.path.join(directories.vardir, "static"))
     sound.media_paths.append(os.path.join(directories.vardir, "assets"))
@@ -32,6 +52,6 @@ def init():
     p = config.config["audio_paths"]
     for i in p:
         if i == "__default__":
-            sound.media_paths.append(os.path.join(directories.datadir, "sounds"))
+            sound.media_paths.append(os.path.join(directories.datadir, "static", "sounds"))
         else:
             sound.media_paths.append(i)